+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+**2026/02/15**
+
+* URL参数新增timeout、followRedirects、resolveIp字段
+* 更换JS引擎，修复const作用域问题
+* 新增原生段评支持
+* java.get/post/head增加重载方法，支持传入json字符串格式的请求头
+* 引入Crypto-js，使阅读原生支持CryptoJS加解密
+* 新增webview代码编辑器，支持自动补全、语法检查、格式化等
+* 新增定时任务功能，使用cron表达式驱动的定时脚本执行管理
+
 **2022/10/02**
 
 * 更新cronet: 106.0.5249.79